requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.3
django-bootstrap-form==3.3
django-embed-video==1.1.2
idna==2.6
nltk==3.2.4
pytz==2017.2
requests==2.18.4
six==1.10.0
urllib3==1.22
virtualenv==15.1.0
